feat(term-setup): First attempt to make multi-term tolerable
I've been wanting to use in-emacs terminal buffers instead of
Alacritty as an EXWM window for a while.

In light of the recent EXWM bugs that cause overlapping X-windows
occasionally I've finally had it!

This commit introduces a new configuration file for multi-term related
settings. Primarily this does the following:

* Introduce a new `C-x t` keybinding which launches a terminal if none
  is running, or prompts the user to select one of the existing ones.
* Remap key bindings in term-mode for quick access to features I want.
* Add a `C-c C-r` key binding in term-mode to quickly rename terminal
  buffers into something sensible. Hopefully this will play nice with
  the ivy-based selector from the first point mentioned here.

I'll see how it goes!